Use json.RawMessage for message data
Currently this package only supports sending of string data. Thus when examining
data in the Satori web consoles you will not be able to explore and expand data
that is sent in JSON format. So to work around the problem we can allow the
publish method to take an interface as its message and we will attempt to
convert the message to the json.RawMessage struct in the golang json package.
This will then allow us to send whatever data type that we want to satori and
guarantee that it can be json marshalled.

Also for testing we can use environment variables that will no longer rely on
hardcoding potential sensitive api-keys and private satori channel endpoints.
This may save the issue of commiting sensitive info by accident and having to
generate new endpoints and api-keys. Also this allows for more configuable
testing to individual channels or for testing the role privledge.